Quotes: Move to a plugin
