+++ server/chat-commands/info.ts
@@ -2536,84 +2536,6 @@ export const commands: ChatCommands = {
 	},
 	denyshowhelp: [`/denyshow [user] - Denies the media display request of [user]. Requires: % @ # &`],
 
-	randquote(target, room, user) {
-		room = this.requireRoom();
-		if (!room.settings.quotes?.length) return this.errorReply(`This room has no quotes.`);
-		this.runBroadcast();
-		const {quote, date, userid} = room.settings.quotes[Math.floor(Math.random() * room.settings.quotes.length)];
-		const time = Chat.toTimestamp(new Date(date), {human: true});
-		const attribution = toID(target) === 'showauthor' ? `<br /><hr /><small>Added by ${userid} on ${time}</small>` : '';
-		return this.sendReplyBox(`${Chat.formatText(quote).replace(/\n/g, '<br />')}${attribution}`);
-	},
-	randquotehelp: [`/randquote [showauthor] - Show a random quote from the room. Add 'showauthor' to see who added it and when.`],
-
-	addquote: 'quote',
-	quote(target, room, user) {
-		room = this.requireRoom();
-		if (!room.persist) {
-			return this.errorReply("This command is unavailable in temporary rooms.");
-		}
-		target = target.trim();
-		this.checkCan('mute', null, room);
-		if (!target) {
-			return this.parse(`/help quote`);
-		}
-		if (!room.settings.quotes) room.settings.quotes = [];
-		if (this.filter(target) !== target) {
-			return this.errorReply(`Invalid quote.`);
-		}
-		if (room.settings.quotes.filter(item => item.quote === target).length) {
-			return this.errorReply(`"${target}" is already quoted in this room.`);
-		}
-		if (target.length > 8192) {
-			return this.errorReply(`Your quote cannot exceed 8192 characters.`);
-		}
-		if (room.settings.quotes.length >= 100) {
-			return this.errorReply(`This room already has 100 quotes, which is the maximum.`);
-		}
-		room.settings.quotes.push({userid: user.id, quote: target, date: Date.now()});
-		room.saveSettings();
-		const collapsedQuote = target.replace(/\n/g, ' ');
-		this.privateModAction(`${user.name} added a new quote: "${collapsedQuote}".`);
-		return this.modlog(`ADDQUOTE`, null, collapsedQuote);
-	},
-	quotehelp: [`/quote [quote] - Adds [quote] to the room's quotes. Requires: % @ # &`],
-
-	removequote(target, room, user) {
-		target = target.trim();
-		const [idx, roomid] = Utils.splitFirst(target, ',');
-		const targetRoom = roomid ? Rooms.search(roomid) : room;
-		if (!targetRoom) return this.errorReply(`Invalid room.`);
-		if (!targetRoom.persist) {
-			return this.errorReply("This command is unavailable in temporary rooms.");
-		}
-		this.room = targetRoom;
-		this.checkCan('mute', null, targetRoom);
-		if (!targetRoom.settings.quotes?.length) return this.errorReply(`This room has no quotes.`);
-		const index = parseInt(idx);
-		if (isNaN(index)) {
-			return this.errorReply(`Invalid index.`);
-		}
-		if (!targetRoom.settings.quotes[index - 1]) {
-			return this.errorReply(`Quote not found.`);
-		}
-		const [removed] = targetRoom.settings.quotes.splice(index - 1, 1);
-		const collapsedQuote = removed.quote.replace(/\n/g, ' ');
-		this.privateModAction(`${user.name} removed quote indexed at ${index}: "${collapsedQuote}" (originally added by ${removed.userid}).`);
-		this.modlog(`REMOVEQUOTE`, null, collapsedQuote);
-		targetRoom.saveSettings();
-		if (roomid) this.parse(`/join view-quotes-${targetRoom.roomid}`);
-	},
-	removequotehelp: [`/removequote [index] - Removes the quote from the room's quotes. Requires: % @ # &`],
-
-	viewquotes: 'quotes',
-	quotes(target, room) {
-		const targetRoom = target ? Rooms.search(target) : room;
-		if (!targetRoom) return this.errorReply(`Invalid room.`);
-		return this.parse(`/join view-quotes-${targetRoom.roomid}`);
-	},
-	quoteshelp: [`/quotes [room] - Shows all quotes for [room]. Defaults the room the command is used in.`],
-
 	approvallog(target, room, user) {
 		room = this.requireRoom();
 		return this.parse(`/sl approved showing media from, ${room.roomid}`);
@@ -2786,33 +2708,6 @@ export const pages: PageTable = {
 			buf += `</div><hr />`;
 		}
 		return buf;
-	},
-	quotes(args, user) {
-		const room = this.requireRoom();
-		this.title = `[Quotes]`;
-		// allow it for users if they can access the room
-		if (!room.checkModjoin(user)) {
-			return this.errorReply(`Access denied.`);
-		}
-		let buffer = `<div class="pad">`;
-		buffer += `<button style="float:right;" class="button" name="send" value="/join view-quotes-${room.roomid}"><i class="fa fa-refresh"></i> Refresh</button>`;
-		if (!room.settings.quotes?.length) {
-			return `${buffer}<h2>This room has no quotes.</h2></div>`;
-		}
-
-		buffer += `<h2>Quotes for ${room.title} (${room.settings.quotes.length}):</h2>`;
-		for (const [i, quoteObj] of room.settings.quotes.entries()) {
-			const index = i + 1;
-			const {quote, userid, date} = quoteObj;
-			buffer += `<div class="infobox">${index}: ${Chat.formatText(quote).replace(/\n/g, '<br />')}`;
-			buffer += `<br /><hr /><small>Added by ${userid} on ${Chat.toTimestamp(new Date(date), {human: true})}</small>`;
-			if (user.can('mute', null, room)) {
-				buffer += `<br /><hr /><button class="button" name="send" value="/removequote ${index},${room.roomid}">Remove</button>`;
-			}
-			buffer += `</div>`;
-		}
-		buffer += `</div>`;
-		return buffer;
 	},
 };
