+++ server/rooms.ts
@@ -111,7 +111,6 @@ export interface RoomSettings {
 	requestShowEnabled?: boolean | null;
 	showEnabled?: GroupSymbol | true;
 	permissions?: {[k: string]: GroupSymbol};
-	quotes?: {userid: string, quote: string, date: number}[];
 
 	scavSettings?: AnyObject;
 	scavQueue?: QueuedHunt[];
